+++ operations.ts
@@ -1,9 +1,17 @@
 import { TypeError } from './exceptions';
 import { RuntimeFault } from './faults';
+import { Value } from './value';
 import { Type } from './value/types';
 import { typeOf } from './value/typeof';
-import { Value } from './value/value';
 import { cast } from './value/cast';
+
+// TODO: I stumbled on a really nasty circular import bug. I was originally
+// importing and re-exporting everything I'm importing here in value's index.
+// When I did that, I triggered the circular import bug.
+//
+// Importing the libraries directly seems to avoid it, and I've moved towards
+// using that pattern more generally. But I suspect that with the wrong
+// imports in any file, I could trigger the issue once again.
 
 function typePrecedence(type: Type): number {
   switch (type) {
