+++ value/index.ts
@@ -1,16 +1,12 @@
-// TODO: Importing and re-exporting everything in here causes a really nasty
-// circular import bug. This can be seen in operations.ts - if it imports
-// from this file, it triggers the bug; but if it imports separately, the
-// issue doesn't occur.
-//
-// For now, I can avoid it by importing libraries directly, and it's probably
-// appropriate to deprecate this import path anyway. But I suspect that any
-// file which imports the wrong combination of these libraries could trigger
-// the issue.
-import { Nil, nil, Value } from './value';
-import { Type } from './types';
-import { typeOf } from './typeof';
-import { cast, castType } from './cast';
-import { into, intoType } from './convert';
+import { BaseException } from '../exceptions';
+import { Formattable, Formatter } from '../format';
 
-export { Nil, nil, Value, Type, typeOf, cast, castType, into, intoType };
+export class Nil implements Formattable {
+  format(_formatter: Formatter): string {
+    return 'nil';
+  }
+}
+
+export const nil = new Nil();
+
+export type Value = number | boolean | string | BaseException | Nil;
